🚚[Rename] 컴포넌트 및 컨테이너 위치 변경
