+++ src/features/chat-room/components/chat-bubble/ChatBubble.stories.tsx
