+++ src/features/chat-room/components/chat-bubble/ChatBubble.tsx
