+++ src/features/chat-room/components/chat-bubble/components/OpponentBubble.tsx
