+++ src/features/chat-room/components/chat-bubble/components/SystemBubble.tsx
