+++ src/features/chat/components/chat-card/types/variantChatCard.ts
